move crypto mod to tests
